Adjust hgrc file when copying test project
The hgrc file contains hard coded paths, so these need to be
adjusted to the current machine before running the tests. Correct
paths in hgrc are required so that Chorus can find e.g.
ChorusMerge.exe. This fixes or at least improves the otherwise
failing SynchronizeActionTest.
