+++ src/LfMerge.Tests/SampleData.cs
@@ -21,7 +21,7 @@ namespace LfMerge.Tests
 			bsonProjectRecordData = BsonSerializer.Deserialize<BsonDocument>(jsonProjectRecordData);
 			bsonOptionListData = BsonSerializer.Deserialize<BsonDocument>(jsonOptionListData);
 			// Semantic domain data is found in $GITROOT/data/semantic-domains/semdom.json
-			string gitRoot = Fdo.FdoTestFixture.FindGitRepoRoot();
+			string gitRoot = TestEnvironment.FindGitRepoRoot();
 			string semDomFilename = Path.Combine(gitRoot, "data", "semantic-domains", "semdom.json");
 			string jsonSemDomData = File.ReadAllText(semDomFilename);
 			bsonSemDomData = BsonSerializer.Deserialize<BsonDocument>(jsonSemDomData);
